requirements.txt
absl-py==2.1.0
accelerate==0.26.1
aiodns==3.1.1
aiofiles==23.1.0
aiohttp==3.8.2
aiosignal==1.3.1
alabaster==0.7.16
alpaca-py==0.10.0
alpaca-trade-api==3.1.1
alpha-vantage==2.3.1
altair==5.0.1
ampligraph==2.1.0
anyio==3.7.0
appdirs==1.4.4
appnope==0.1.4
APScheduler==3.10.4
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.2
beautifultable==1.1.0
bidict==0.22.1
bleach==6.2.0
blinker==1.6.2
blis==0.7.11
boto3==1.34.131
botocore==1.34.131
branca==0.8.1
cachetools==4.2.4
cairocffi==1.7.1
Cartopy==0.23.0
catalogue==2.0.10
ccxt==3.0.61
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cloudpathlib==0.16.0
coloredlogs==14.3
comm==0.2.2
confection==0.1.4
contextlib2==21.6.0
contourpy==1.0.7
cryptography==42.0.2
cycler==0.11.0
cymem==2.0.8
datasketch==1.6.5
dateparser==1.2.0
debugpy==1.8.2
decorator==4.4.2
defusedxml==0.7.1
deprecation==2.1.0
diffusers==0.26.2
Django==5.0.2
dnspython==2.5.0
docopt==0.6.2
docutils==0.17.1
docx==0.2.4
docx2txt==0.8
editdistance==0.8.1
email-validator==2.1.0.post1
et-xmlfile==1.1.0
Events==0.5
exchange_calendars==4.5.2
executing==2.0.1
fastapi==0.96.0
fastjsonschema==2.21.1
fasttext==0.9.3
ffmpy==0.3.0
filelock==3.12.0
fiona==1.10.1
flake8==7.1.1
Flask==2.3.2
Flask-BabelEx==0.9.4
Flask-Login==0.6.3
Flask-Mail==0.9.1
flask-marshmallow==1.2.0
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flatbuffers==24.3.25
folium==0.15.1
fonttools==4.39.4
fpdf==1.7.2
frozendict==2.4.0
frozenlist==1.3.3
fs==2.4.16
fsspec==2023.5.0
gast==0.6.0
gensim==4.3.3
geographiclib==2.0
geopandas==0.14.1
gitdb==4.0.11
GitPython==3.1.41
google-api-core==2.16.2
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.4.1
google-cloud-translate==3.15.0
google-pasta==0.2.0
googleapis-common-protos==1.62.0
gradio==3.33.1
gradio_client==0.2.5
grpcio==1.60.1
grpcio-status==1.60.1
gspread==5.10.0
h11==0.14.0
h3==3.7.7
h5py==3.11.0
html5lib==1.1
httpcore==0.17.2
httplib2==0.22.0
httpx==0.24.1
huggingface-hub==0.20.3
humanfriendly==10.0
ibapi==9.81.1.post1
ibm-cloud-sdk-core==3.16.7
ibm-watson==7.0.0
idna==3.4
imageio==2.31.0
imageio-ffmpeg==0.4.8
imagesize==1.4.1
importlib-metadata==7.0.1
inflection==0.5.1
iniconfig==2.0.0
instaloader @ git+https://github.com/instaloader/instaloader@0ff91cceb5e8c17b545ec020b2b33a429b712aa8
instalooter==2.4.4
invisible-watermark==0.2.0
ipykernel==6.29.5
ipython==8.12.3
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonpickle==3.0.2
jsonschema==4.17.3
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keras==3.4.1
kiwisolver==1.4.4
korean-lunar-calendar==0.3.1
langcodes==3.3.0
latexcodec==3.0.0
libclang==18.1.1
libpypostal==0.0.6
linkify-it-py==2.0.2
lumibot==2.8.11
lxml==4.9.3
Markdown==3.6
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.20.2
marshmallow-sqlalchemy==1.0.0
matplotlib==3.8.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==3.1.3
ml-dtypes==0.4.0
mmh3==4.1.0
more-itertools==10.2.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.3
multidict==5.2.0
multitasking==0.0.11
murmurhash==1.0.10
myst-parser==0.18.0
namex==0.0.8
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
neo4j==5.21.0
nest-asyncio==1.6.0
networkx==3.2
newsapi-python==0.2.7
nltk==3.8.1
numpy==1.24.3
nvector==0.7.7
oauth2client==4.1.3
oauthlib==3.2.2
odfpy==1.4.1
openai==0.27.7
opencv-python==4.9.0.80
openpyxl==3.1.5
opensearch==0.9.2
opensearch-py==2.6.0
opt-einsum==3.3.0
optree==0.12.1
orjson==3.9.0
outcome==1.3.0.post0
packaging==23.1
pandas==2.0.3
pandas-datareader==0.10.0
pandas_market_calendars==4.4.0
pandocfilters==1.5.1
parso==0.8.3
passlib==1.7.4
pdfkit==1.0.0
pdfminer.six==20231228
peewee==3.17.1
pexpect==4.9.0
pi==0.1.2
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.5.0
platformdirs==4.2.2
plotly==5.18.0
pluggy==1.4.0
polygon-api-client==1.13.4
preshed==3.0.9
proglog==0.1.10
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py3-wkhtmltopdf==0.4.1
pyamaze==1.0.1
pyarrow==15.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.13.1
pybtex==0.24.0
pybtex-docutils==1.0.3
pycares==4.4.0
pycodestyle==2.12.1
pycountry==22.3.5
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.8
pydeck==0.8.1b0
pydub==0.25.1
pyflakes==3.2.0
pygame==2.6.1
Pygments==2.15.1
PyJWT==2.8.0
pyluach==2.2.0
pymaze==1.2.3
pymongo==4.13.0
PyNaCl==1.5.0
pyogrio==0.9.0
pyparsing==3.0.9
PyPDF2==3.0.1
pyproj==3.6.1
pyresparser==1.0.6
pyrsistent==0.19.3
pyshp==2.3.1
PySocks==1.7.1
pytest==8.0.0
python-calamine==0.2.3
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.1
python-engineio==4.9.0
python-multipart==0.0.6
python-socketio==5.11.1
pytube==15.0.0
pytz==2023.3
PyWavelets==1.5.0
PyYAML==6.0.1
pyzmq==26.0.3
qddate==1.0.4
Quandl==3.7.0
QuantStats==0.0.62
rdflib==7.0.0
regex==2023.10.3
replicate==0.23.1
reportlab==4.0.4
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
s3transfer==0.10.1
safetensors==0.4.2
schedule==0.6.0
schema==0.7.5
scikit-learn==1.3.1
scipy==1.10.0
seaborn==0.13.2
selenium==4.18.1
semantic-version==2.10.0
shapely==2.0.4
simple-websocket==1.0.0
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
speaklater==1.3
SpeechRecognition==3.10.0
Sphinx==5.0.2
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-bibtex==2.4.2
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.25
sqlparse==0.4.4
srsly==2.4.8
sseclient-py==1.8.0
stack-data==0.6.3
starlette==0.27.0
streamlit==1.31.0
sumolib==1.23.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
thinc==8.2.3
threadpoolctl==3.2.0
timedelta==2020.12.3
tinycss2==1.4.0
tkcalendar==1.6.1
tokenizers==0.15.1
toml==0.10.2
toolz==0.12.0
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
tornado==6.4
tqdm==4.65.0
traci==1.18.0
traitlets==5.14.1
transformers==4.37.2
trio==0.25.0
trio-websocket==0.11.1
typer==0.9.4
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
uc-micro-py==1.0.2
uritemplate==3.0.1
urllib3==2.2.2
uvicorn==0.22.0
validators==0.22.0
verboselogs==1.7
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
webencodings==0.5.1
websocket-client==1.1.0
websockets==10.4
Werkzeug==2.3.6
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.1.2
xyzservices==2025.4.0
yarg==0.1.9
yarl==1.9.2
yfinance==0.2.36
youtube-dl==2021.12.17
zipp==3.17.0
